+++ CHANGELOG.rst
@@ -8,6 +8,11 @@ Version 4.1.0
 - Added method ``manufacturer()`` for class ``Transport()``
 - Added ``sk`` (Slovak) locale support
 
+**Updated**:
+
+- Update parameter ``end`` for some methods of provider ``Datetime()`` (Fix #870)
+
+
 Version 4.0.0
 -------------