+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ Version 4.1.0
 
 **Removed**:
 
-- Removed the deprecated ``download_image()`` function from the ``shortcuts`` module, use your custom downloader instead.
+- Removed the deprecated ``download_image()`` function from the ``shortcuts`` module, use your own custom downloader instead.
 
 Version 4.0.0
 -------------